+++ database/factories/ModelFactory.php
@@ -1,7 +1,7 @@
 <?php
 
 use Faker\Generator as Faker;
-use Thinktomorrow\AssetLibrary\Test\stubs\Article;
+use Thinktomorrow\AssetLibrary\Tests\stubs\Article;
 
 /*
 |--------------------------------------------------------------------------