+++ src/Utilities/AssetFormHelper.php
@@ -8,7 +8,7 @@ class AssetFormHelper
      * Generates the hidden field that links the file to a specific type.
      *
      * @param string $type
-     * @param null $locale
+     * @param string|null $locale
      *
      * @param string $name
      * @return string