+++ tests/TestHelpers.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Thinktomorrow\AssetLibrary\Test;
+namespace Thinktomorrow\AssetLibrary\Tests;
 
 use Spatie\MediaLibrary\Models\Media;
 use Thinktomorrow\AssetLibrary\Models\Asset;
-use Thinktomorrow\AssetLibrary\Test\stubs\Article;
+use Thinktomorrow\AssetLibrary\Tests\stubs\Article;
 
 trait TestHelpers
 {