+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All Notable changes to `AssetLibrary` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## NEXT - 2018-11-06
+
+### Added
+- Added HasAsset interface to replace the HasMedia interface on the models.
+
 ## NEXT - 2018-07-26
 
 ### Added